% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_BP.R
\name{family_BP}
\alias{family_BP}
\alias{BloodPressure}
\title{Blood Pressure Assessment Functions for Paediatric and Adult Populations}
\description{
This module provides functions to assess blood pressure (BP) in children,
young people, and adults using age-appropriate clinical guidelines, currently
the package uses:

\itemize{
  \item Children/adolescents (0-17 years): NHBPEP Fourth Report
  \item Young adults (>17 years): NICE/BHF guidelines
}
}
\details{
The functions implement clinical BP categorisation following:
\itemize{
  \item **NHBPEP Fourth Report**: Sex-, age-, and height-specific percentiles for paediatric BP
  \item **NICE/BHF Guidelines**: Fixed thresholds for adult hypertension stages
  \item **NPDA Limits**: National Paediatric Diabetes Audit validity ranges (50/15-200/150 mmHg)
  \item **NDA Limits**: National Diabetes Audit validity ranges (70/20-300/150 mmHg)
}
}
\section{BP Categories by Age Group}{

\tabular{lllc}{
  \strong{Category} \tab \strong{Age (years)} \tab \strong{Reference} \tab \strong{Criteria} \cr
  Normotension \tab [0-12] (children)      \tab NHBPEP Fourth Report (NPDA min)  \tab (50/15mmHg ≤) SBP/DBP < 90th percentile \cr
  Normotension \tab (12-17] (adolescents)  \tab NHBPEP Fourth Report (NPDA min)  \tab (50/15mmHg ≤) SBP/DBP < 120/80 mmHg \cr
  Normotension \tab >17                    \tab NICE/BHF (NDA min)               \tab (70/20mmHg ≤) SBP/DBP < 120/80 mmHg \cr
  Prehypertension \tab [0-12] (children)     \tab NHBPEP Fourth Report             \tab 90th ≤ SBP/DBP < 95th \cr
  Prehypertension \tab (12-17] (adolescents) \tab NHBPEP Fourth Report             \tab 120/80mmHg ≤ SBP/DBP < 95th \cr
  Prehypertension \tab >17                   \tab NICE/BHF                         \tab 120/80mmHg ≤ SBP/DBP < 140/90mmHg \cr
  Stage 1 hypertension \tab [0-1) (newborns) \tab NHBPEP Fourth Report             \tab 95th ≤ SBP < 5mmg above 99th \cr
  Stage 1 hypertension \tab [1-17]           \tab NHBPEP Fourth Report             \tab 95th ≤ SBP/DBP < 5mmg above 99th \cr
  Stage 1 hypertension \tab >17              \tab NICE/BHF                         \tab 140/90mmHg ≤ SBP/DBP < 160/100mmHg \cr
  Stage 2 hypertension \tab [0-1) (newborns) \tab NHBPEP Fourth Report (NPDA max) \tab 5mmg above 99th ≤ SBP (≤ 200/150mmHg) \cr
  Stage 2 hypertension \tab [1-17]           \tab NHBPEP Fourth Report (NPDA max) \tab 5mmg above 99th ≤ SBP/DBP (≤ 200/150mmHg) \cr
  Stage 2 hypertension \tab >17              \tab NICE/BHF                          \tab 160/100mmHg ≤ SBP/DBP < 180/120mmHg \cr
  Stage 3 hypertension \tab >17              \tab NICE/BHF (NDA max)              \tab 180/120mmHg ≤ SBP/DBP (≤ 300/150mmHg) \cr
}
}

\references{
\itemize{
  \item NHBPEP Fourth Report: \href{https://www.nhlbi.nih.gov/files/docs/resources/heart/hbp_ped.pdf}{Paediatric BP categories}
  \item NICE Guidelines: \href{https://cks.nice.org.uk/topics/hypertension/}{Adult BP categories}
  \item British Heart Foundation: \href{https://www.bhf.org.uk/informationsupport/risk-factors/high-blood-pressure}{Adult BP categories}
  \item NDA Methodology: \href{https://digital.nhs.uk/data-and-information/publications/statistical/national-diabetes-audit/core-report-1-2020-21/methodology}{Adult BP Range}
  \item NPDA Methodology: \href{https://www.rcpch.ac.uk/work-we-do/clinical-audits/npda/transparency-open-data}{Paediatric BP Range}
  \item RCPCH Growth API: \href{https://growth.rcpch.ac.uk/}{Python library for Paediatric SDS calculation}
  \item childsds package: \href{https://mvogel78.r-universe.dev/childsds}{R package for Paediatric SDS calculation}
}
}
\seealso{
Other BP functions: 
\code{\link{get_BP}()},
\code{\link{get_BPCategory}()},
\code{\link{get_BPExpected}()},
\code{\link{get_BPRelative}()}
}
\author{
Zhaonan Fang
}
\concept{BP functions}
